// web/src/api/generated/model/moduleReadingList.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { ListedBook } from './listedBook';

export interface ModuleReadingList {
  offering_id: string;
  module_code: string;
  module_name: string;
  class_group: string | null;
  books: ListedBook[];
}